pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "KubeZen"
version = "0.1.0"
description = "A Kubernetes CLI tool"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
  {name = "Georgi Panov", email = "georgi.panov@darkfellanetwork.com"},
]
dependencies = [
    "click",
    "libtmux",
    "rich",
    "requests",
    "kubernetes",
    "PyYAML",
    "kubernetes_asyncio",
    "aiofiles",
    "typing-extensions",
]

[project.scripts]
kubezen = "KubeZen.cli:cli"

[tool.black]
line-length = 99
